requirement.txt
appdirs==1.4.4
bcrypt==3.1.7
blinker==1.6.3
cffi==1.16.0
click==8.1.7
configparser==6.0.0
cryptography==2.8
Fabric3==1.14.post1
Flask==3.0.0
Flask-Cors==4.0.0
greenlet==3.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
paramiko==2.12.0
protobuf==4.21.12
pycparser==2.21
PyNaCl==1.3.0
pyparsing==3.1.1
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==3.0.0
zipp==3.17.0
